fix: 絶対パスを`src/hoge/hoge`のようにsrcから始まるパスに変更して、動作を確認した
