--- src/component.md
@@ -1,21 +1,0 @@
-# 概要
-
-作る前に現状のサイトの DOM を理解する。
-
-その上で「このようなパーツで分けて作成する」という指針
-
-## 構成
-
-- base
-  - header
-  - main
-    - hero_img
-    - portfolio
-    - about_me
-    - and_more
-  - footer
-
-## 補足
-
-- header や footer は要素が少ないので随時修正する
-- 現状は 8 個のコンポーネントを作成予定
