+++ src/scripts/main/aboutMe/AboutMeWork.js
@@ -1,6 +1,6 @@
 /** @jsx jsx */
 import { css, jsx } from "@emotion/core";
-import { baseCSS } from "Const";
+import { baseCSS } from "src/scripts/Const";
 
 const profileListDetails = css`
   color: ${baseCSS.titleColor};
